Turn off strict parsing for pls playlist files
Some PLS files contain one 'Version' key for each file in the playlist.
This ultimately has no impact on how mopidy parses files in the
playlist, and therefore it should be as tolerant as possible of
real-world playlist files.

This matches a similar change request submitted to the mopidy core:

https://github.com/mopidy/mopidy/pull/1923